cleanup and separate python/node
