+++ .gitignore
@@ -1,5 +1,5 @@
 .DS_Store
-__pycache__
+python/__pycache__
 
 # Logs
 logs
