+++ node/package.json
@@ -12,10 +12,10 @@
     "amqplib": "^0.5.2",
     "body-parser": "~1.18.2",
     "cookie-parser": "~1.4.3",
-    "csvtojson": "^1.1.9",
     "debug": "~2.6.9",
     "express": "~4.15.5",
     "express-validator": "^4.3.0",
+    "http-status-codes": "^1.3.0",
     "morgan": "~1.9.0",
     "node-sass-middleware": "0.9.8",
     "pug": "2.0.0-beta11",
