+++ node/public/stylesheets/style.sass
@@ -1,0 +1,6 @@
+body
+  padding: 50px
+  font: 14px "Lucida Grande", Helvetica, Arial, sans-serif
+
+a
+  color: #00B7FF
